+++ packages/create-vuestic/src/steps/2.addVuestic/insert-vuestic-compiler.ts
